Got the very first part of the base page working again, including ensuring that streamlit caching works okay.
